Simplify reduction into a single method in ManagedArray.h.
